backend/requirements.txt
Django==4.2
djangorestframework
pandas
reportlab
django-cors-headers
gunicorn
python-dotenv
